cambio apiResponse y comentarios
cambio apiResponsen OFICIO-TRABAJOAPP-ROL y comentarios

+++ src/main/java/com/findfix/find_fix_app/oficio/service/OficioServiceImpl.java
@@ -40,7 +40,11 @@ public class OficioServiceImpl implements OficioService {
     }
 
     @Override
-    public void borrarOficioPorId(Long id) {
+    public void borrarOficioPorId(Long id) throws OficioNotFoundException {
+        if(oficioRepository.findById(id).isEmpty())
+        {
+            throw new OficioNotFoundException("El id ingresado no pertenece a ningun oficio registrado");
+        }
         oficioRepository.deleteById(id);
     }
